refactor(filters): relocate generic mongodb filter modules to wandb/_filters
Pure relocation of the generic MongoDB filter modules (expressions, operators, filterutils, and the package initializer) out of wandb/automations/_filters into a new top-level private package wandb/_filters. No behavior change and no backward-compatibility shims. Every importer is updated directly to point at the new location.

The automations-specific modules run_metrics and run_states stay in wandb/automations/_filters and now import their generic siblings from wandb/_filters.

Part of a stack to enable reusing these filters for paginator validation.
