+++ wandb/_filters/__init__.py
@@ -1,9 +1,7 @@
-from .expressions import FilterableField, FilterExpr, MongoLikeFilter
-from .filterutils import simplify_expr
+from .expressions import FilterExpr, MongoLikeFilter
 from .operators import (
     All,
     And,
-    BaseOp,
     Contains,
     Eq,
     Exists,
@@ -23,7 +21,6 @@ from .operators import (
 )
 
 __all__ = [
-    "BaseOp",
     "And",
     "Or",
     "Nor",
@@ -42,8 +39,6 @@ __all__ = [
     "Regex",
     "Size",
     "All",
-    "FilterableField",
     "FilterExpr",
     "MongoLikeFilter",
-    "simplify_expr",
 ]
